pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentlab"
dynamic = ["version"]
description = "Main package for developing agents and experiments"
authors = [
    {name = "Rim Assouel", email = "rim.assouel@gmail.com"},
    {name = "Alex Drouin", email = "alex.drouin@servicenow.com"},
    {name = "Maxime Gasse", email = "maxime.gasse@servicenow.com"},
    {name = "Alex Lacoste", email = "alex.lacoste@servicenow.com"},
    {name = "Tom Marty", email = "tom.marty@polymtl.ca"},
    {name = "Massimo Caccia", email = "massimo.caccia1@servicenow.com"},
    {name = "Thibault Le Sellier de Chezelles", email = "thibault.de.chezelles@gmail.com"},
    {name = "Aman Jaiswal", email = "aman.jaiswal@servicenow.com"},
]
readme = "README.md"
requires-python = ">3.10"
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "pydantic~=2.9",
    "dask",
    "distributed",
    "browsergym>=0.7.1",
    "joblib>=1.2.0",
    "openai>=1.7,<2",
    "tiktoken",
    "huggingface_hub",
    "contexttimer",
    "ipython",
    "pyyaml>=6",
    "pandas",
    "gradio>=5.5",
    "gitpython",
    "requests",
    "matplotlib",
    "ray[default]",
    "python-slugify",
    "pillow",
    "gymnasium>=0.27",
    "torch>=2.2.2",
    "safetensors>=0.4.0",
    "transformers>=4.38.2",
    "anthropic>=0.62.0",
    "litellm>=1.75.3",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "black[jupyter]>=24.2.0",
    "blacken-docs",
    "pre-commit",
    "pytest==7.3.2",
    "flaky",
    "pytest-xdist",
    "pytest-playwright",
]
# tapeagents = [
#     "tapeagents[converters]",
# ]

[project.urls]
"Homepage" = "https://github.com/ServiceNow/AgentLab"

[tool.setuptools.dynamic]
version = {attr = "agentlab.__version__"}

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.nox
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[dependency-groups]
dev = [
    "darglint>=1.8.1",
    "ipykernel>=6.30.1",
    "pip>=25.2",
    "pytest-cov>=7.0.0",
]
hint = [
    "sentence-transformers>=5.0.0",
]


[project.scripts]
agentlab-assistant = "agentlab.ui_assistant:main"
agentlab-xray = "agentlab.analyze.agent_xray:main"
agentlab-mentor = "agentlab.agents.hitl_agent.launch_hint_ui:main"
